% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdiag.R
\name{isdiag}
\alias{isdiag}
\title{Diagnostics for importance sampling from stanoptim}
\usage{
isdiag(fit, wait = TRUE)
}
\arguments{
\item{fit}{Output from stanoptim when isloops > 0}
}
\value{
Nothing. Plots convergence of parameter mean estimates from initial Hessian based distribution to final sampling distribution.
}
\description{
Diagnostics for importance sampling from stanoptim
}
\examples{
\dontrun{
library(rstan)
scode <- "
parameters {
  real y[2];
}
model {
  y[1] ~ normal(0, 1);
  y[2] ~ double_exponential(0, 2);
}
"

sm <- stan_model(model_code=scode)

fit1 <- sampling(object = sm, iter = 10, verbose = FALSE)
print(fit1)
fit2 <- stan(fit = fit1, iter = 10000, verbose = FALSE)

## extract samples as a list of arrays
e2 <- extract(fit2, permuted = TRUE)

## using as.array on the stanfit object to get samples
a2 <- as.array(fit2)


optimfit <- optimstan(standata = list(),sm = sm,isloops=10,issamples = 1000,cores=3)

apply(optimfit$rawposterior,2,mean)
apply(optimfit$rawposterior,2,sd)
isdiag(optimfit)

plot(density(optimfit$rawposterior))
points(density(e2$y))
}
}
