% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimstan.R
\name{optimstan}
\alias{optimstan}
\title{Optimize / importance sample a stan or ctStan model.}
\usage{
optimstan(standata, sm, init = 0, deoptim = FALSE, decontrol = list(),
  isloops = 5, isloopsize = 500, issamples = 500, cores = 1)
}
\arguments{
\item{standata}{list object conforming to rstan data standards.}

\item{sm}{compiled stan model object.}

\item{init}{init argument conforming to rstan init standards.}

\item{deoptim}{Do first pass optimization using differential evolution? Slower, but better for cases with multiple
minima / difficult optimization.}

\item{decontrol}{List of control parameters for differential evolution step, to pass to \code{\link[DEoptim]{DEoptim.control}}.}

\item{isloops}{Only relevent if \code{optimize=TRUE}.
Number of iterations of adaptive importance sampling to perform after optimization.}

\item{isloopsize}{Only relevent if \code{optimize=TRUE}.
Number of samples per iteration of importance sampling.}

\item{issamples}{Number of samples to use for final results of importance sampling.}

\item{cores}{Number of cpu cores to use.}

\item{verbose}{Integer from 0 to 2. Higher values print more information during model fit -- for debugging.}

\item{nopriors}{logical. If TRUE, any priors are disabled -- sometimes desirable for optimization.}

\item{ukfspread}{Numeric governing the spread of sigma points when using the unscented Kalman filter. Smaller values (e.g. 1e-2)
are apparently typical, but this doesn't seem sensible to me.}
}
\value{
list containing: optimfit, the result from optimizing with stans optimizer;
stanfit, the stanfit object used;
posterior, the sampled posterior;
transformedpars, the full model output based on the posterior;
transformedpars_old, a simple summary approach based on the minimum and Hessian;
isdiags, a list containing elements used by \code{\link{isdiag}} for importance sampling diagnostics.
}
\description{
Optimize / importance sample a stan or ctStan model.
}
\examples{
#' \dontrun{
library(rstan)
scode <- "
parameters {
  real y[2];
}
model {
  y[1] ~ normal(0, 1);
  y[2] ~ double_exponential(0, 2);
}
"

sm <- stan_model(model_code=scode)
fit <- stan(fit = fit1, iter = 10000, verbose = FALSE)

## extract samples as a list of arrays
e <- extract(fit, permuted = TRUE)

optimfit <- optimstan(standata = list(),sm = sm,isloops=10,issamples = 1000,cores=3)

apply(optimfit$rawposterior,2,mean)
apply(optimfit$rawposterior,2,sd)
isdiag(optimfit)

plot(density(optimfit$rawposterior))
points(density(e2$y))
}
}
